Add Easy Java Solution from LeetCode

+++ LeetCode/Java/Easy/989. Add to Array-Form of Integer/Solution.java
@@ -1,7 +1,7 @@
 class Solution {
     public List<Integer> addToArrayForm(int[] num, int k) {
-        int n=0;
-        int m=1;
+        long n=0;
+        long m=1;
         for(int i=num.length-1;i>=0;i--){
             n +=num[i]*m;
             m=m*10;
@@ -11,8 +11,8 @@ class Solution {
 
         List<Integer> la=new ArrayList<>();
         while(n!=0){
-            int t=n%10;
-            la.addFirst(t);
+            long t=n%10;
+            la.addFirst((int)t);
             n=n/10;
         }
         return la;
